pyproject.toml
[project]
name = "scanovich-content"
version = "0.1.0"
description = "Telegram bot for marketplace product image generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0",
    "httpx>=0.25",
    "openai>=1.0",
    "google-genai>=0.2.0",
    "redis>=5.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/bot", "src/services", "src/utils"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"

